Combat system completed.
Changed the code around a little to fix a few combat and movement control bugs. Added "projectile" for the long range - a ray, as implementing actual projectiles at this stage would be a real pain and not really worth the effort. Renamed characters to match our project and added labels to make it easier to differ between the characters.